feat(next/web): auto scroll into view
